Fix compiler warning on x86_64.
git-svn-id: file:///.data/svn_convert/svn@6737 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/0.9.4/plugins/ontop/ontop.cpp
@@ -99,7 +99,7 @@ OnTopPlugin::OnTopPlugin(unsigned base, Buffer *config)
 
 OnTopPlugin::~OnTopPlugin()
 {
-    Event eCmd(EventCommandRemove, (void*)CmdOnTop);
+    Event eCmd(EventCommandRemove, (void*)(long)CmdOnTop);
     eCmd.process();
 
     setOnTop(false);
